+++ ppci/runtime.py
@@ -65,7 +65,7 @@ module bsp;
 
 public function void putc(byte c)
 {
-  syscall(1, 1, cast<int>(&c), 1);
+  syscall(1, 1, cast<int64_t>(&c), 1);
 }
 
 function void exit()
@@ -73,7 +73,7 @@ function void exit()
     syscall(60, 0, 0, 0);
 }
 
-function void syscall(int nr, int a, int b, int c);
+function void syscall(int64_t nr, int64_t a, int64_t b, int64_t c);
 
 """
